+++ src/config/networks.config.ts
@@ -2,6 +2,13 @@ import { BinanceSmartChain } from 'mobx/model/network/bsc.network';
 import { Ethereum } from 'mobx/model/network/eth.network';
 import { Polygon } from 'mobx/model/network/matic.network';
 import { Network } from 'mobx/model/network/network';
+import { xDai } from 'mobx/model/network/xdai.network';
+import { FLAGS } from './constants';
 
-export const supportedNetworks: Network[] = [new Ethereum(), new Polygon(), new BinanceSmartChain()];
+export const supportedNetworks: Network[] = [
+	new Ethereum(),
+	new BinanceSmartChain(),
+	...(FLAGS.MATIC ? [new Polygon()] : []),
+	...(FLAGS.XDAI ? [new xDai()] : []),
+];
 export const defaultNetwork = supportedNetworks[0];