+++ src/mobx/RootStore.ts
@@ -12,7 +12,6 @@ import { NETWORK_IDS } from '../config/constants';
 import { HoneyPotStore } from './stores/honeyPotStore';
 import UserStore from './stores/UserStore';
 import { LeaderBoardStore } from './stores/LeaderBoardStore';
-import { BoostOptimizerStore } from './stores/boostOptimizerStore';
 import PricesStore from './stores/PricesStore';
 import { NetworkStore } from './stores/NetworkStore';
 
@@ -32,7 +31,6 @@ export class RootStore {
 	public honeyPot: HoneyPotStore;
 	public user: UserStore;
 	public leaderBoard: LeaderBoardStore;
-	public boostOptimizer: BoostOptimizerStore;
 	public prices: PricesStore;
 
 	constructor() {
@@ -51,7 +49,6 @@ export class RootStore {
 		this.setts = new SettStore(this);
 		this.user = new UserStore(this);
 		this.leaderBoard = new LeaderBoardStore(this);
-		this.boostOptimizer = new BoostOptimizerStore(this);
 		this.prices = new PricesStore(this);
 	}