+++ src/mobx/model/contract/batch-response.ts
@@ -1,0 +1,10 @@
+export interface BatchResponse {
+	address: string;
+	namespace: string;
+}
+
+export interface BatchContractItem {
+	args: string[];
+	input: string;
+	value: string;
+}