+++ src/mobx/stores/rebaseStore.ts
@@ -4,11 +4,15 @@ import BatchCall from 'web3-batch-call';
 import { RootStore } from '../RootStore';
 import { getNextRebase, getRebaseLogs } from '../utils/diggHelpers';
 import { RebaseInfo } from 'mobx/model/tokens/rebase-info';
+import DroptRedemption from '../../config/system/abis/DroptRedemption.json';
+import { AbiItem } from 'web3-utils';
+import { getSendOptions } from 'mobx/utils/web3';
 import { ProviderReport } from 'mobx/model/digg/provider-reports';
 import { OracleReports } from 'mobx/model/digg/oracle';
 import { getRebase } from 'config/system/rebase';
 import BigNumber from 'bignumber.js';
 import { groupBy } from 'utils/lodashToNative';
+import { DroptContractResponse } from 'mobx/model/tokens/dropt-info';
 
 let batchCall: any = null;
 
@@ -50,11 +54,28 @@ class RebaseStore {
 
 		const diggData = await batchCall.execute(rebaseConfig.digg);
 		const keyedResult = groupBy(diggData, (v) => v.namespace);
-		const { policy, token, oracle } = keyedResult;
+		const { policy, token, oracle, dropt } = keyedResult;
 
 		if (!this.hasCallResults(token) || !this.hasCallResults(policy) || !this.hasCallResults(oracle)) {
 			return;
 		}
+
+		// dropt data
+		const validDropts = dropt
+			.filter(
+				(_dropt: DroptContractResponse) =>
+					Number(_dropt.expirationTimestamp[0].value) > Number(_dropt.getCurrentTime[0].value) &&
+					Number(_dropt.expiryPrice[0].value) > 0,
+			)
+			.map((validDropt: DroptContractResponse) => {
+				return {
+					[validDropt['address']]: {
+						expiryPrice: validDropt.expiryPrice[0].value,
+						expirationTimestamp: validDropt.expirationTimestamp[0].value,
+						currentTimestamp: validDropt.getCurrentTime[0].value,
+					},
+				};
+			});
 
 		// policy data
 		const latestRebase = Number(policy[0].lastRebaseTimestampSec[0].value);
@@ -86,11 +107,42 @@ class RebaseStore {
 			oracleRate: new BigNumber(activeReport.value.payload).dividedBy(1e18),
 			nextRebase: getNextRebase(minRebaseInterval, latestRebase),
 			pastRebase: rebaseLog,
+			validDropts: validDropts,
 		};
 	});
 
 	private hasCallResults(results: any[]): boolean {
 		return !!results && results.length > 0;
+	}
+
+	public async redeemDropt(redemptionContract: string, redeemAmount: BigNumber): Promise<void> {
+		if (redeemAmount.lte(0)) {
+			return;
+		}
+		const { queueNotification, gasPrice } = this.store.uiState;
+		const { provider, connectedAddress } = this.store.wallet;
+		const { gasPrices } = this.store.network;
+
+		const web3 = new Web3(provider);
+		const redemption = new web3.eth.Contract(DroptRedemption.abi as AbiItem[], redemptionContract);
+		const method = redemption.methods.redeem(redeemAmount);
+
+		queueNotification(`Sign the transaction to claim your options`, 'info');
+
+		const price = gasPrices[gasPrice];
+		const options = await getSendOptions(method, connectedAddress, price);
+		await method
+			.send(options)
+			.on('transactionHash', (_hash: string) => {
+				queueNotification(`Claim submitted.`, 'info', _hash);
+			})
+			.on('receipt', () => {
+				queueNotification(`Options claimed.`, 'success');
+				this.store.user.updateBalances();
+			})
+			.on('error', (error: Error) => {
+				queueNotification(error.message, 'error');
+			});
 	}
 }