+++ src/pages/Landing.tsx
@@ -12,6 +12,7 @@ import SettList from 'components-v2/landing/SettList';
 import { RewardsModal } from '../components-v2/landing/RewardsModal';
 import { SettState } from '../mobx/model/setts/sett-state';
 import { HeaderContainer, LayoutContainer } from '../components-v2/common/Containers';
+import { FLAGS } from 'config/constants';
 
 const useStyles = makeStyles((theme) => ({
 	marginTop: {
@@ -65,6 +66,9 @@ const useStyles = makeStyles((theme) => ({
 	announcementButton: {
 		marginTop: theme.spacing(3),
 		pointerEvents: 'none',
+	},
+	linkButton: {
+		marginTop: theme.spacing(3),
 	},
 }));
 
@@ -123,10 +127,23 @@ const Landing = observer((props: LandingProps) => {
 				</Grid>
 			</Grid>
 
-			{state !== SettState.Open && (
+			{state === SettState.Guarded && (
 				<Grid container spacing={1} justify="center">
 					<Button className={classes.announcementButton} size="small" variant="outlined" color="primary">
 						Note: New Vaults may take up to 2 weeks from launch to reach full efficiency.
+					</Button>
+				</Grid>
+			)}
+			{state === SettState.Open && FLAGS.BOOST_OPTIMIZER && (
+				<Grid container spacing={1} justify="center">
+					<Button
+						className={classes.linkButton}
+						size="small"
+						variant="contained"
+						color="primary"
+						onClick={() => window.open('https://badger.wiki/badger-boost')}
+					>
+						Badger Boost V2 Has Launched - Click Here for more information
 					</Button>
 				</Grid>
 			)}