+++ src/utils/componentHelpers.ts
@@ -1,6 +1,4 @@
-import { LEADERBOARD_RANKS } from '../config/constants';
-import { isWithinRange } from '../mobx/utils/helpers';
-import { LeaderboardRank } from '../mobx/model/boost/leaderboard-rank';
+export const restrictToRange = (num: number, min: number, max: number): number => Math.min(Math.max(num, min), max);
 
 export const debounce = (n: number, fn: (...params: any[]) => any, immediate = false): any => {
 	let timer: any = undefined;
@@ -23,33 +21,17 @@ export const shortenAddress = (address: string): string => {
  * Calculates the percentage of a given point within a range
  */
 export const percentageBetweenRange = (point: number, upperLimit: number, lowerLimit: number): number => {
+	if (point < lowerLimit) {
+		return 0;
+	}
+
+	if (point >= upperLimit) {
+		return 100;
+	}
 	return ((point - lowerLimit) / (upperLimit - lowerLimit)) * 100;
 };
 
-export const getRankFromBoost = (currentBoost: number): LeaderboardRank => {
-	return LEADERBOARD_RANKS[getRankNumberFromBoost(currentBoost)];
-};
-
-export const getRankNumberFromBoost = (currentBoost: number): number => {
-	if (currentBoost < LEADERBOARD_RANKS[LEADERBOARD_RANKS.length - 1].boostRangeStart) {
-		return LEADERBOARD_RANKS.length - 1;
-	}
-
-	// ranks are in descending order
-	for (let index = LEADERBOARD_RANKS.length - 1; index >= 0; index--) {
-		const currentBadgerLevel = LEADERBOARD_RANKS[index];
-		const nextBadgerLevel = LEADERBOARD_RANKS[index - 1];
-
-		// boost has reached last level
-		if (!nextBadgerLevel) {
-			return index;
-		}
-
-		if (isWithinRange(currentBoost, currentBadgerLevel.boostRangeStart, currentBadgerLevel.boostRangeEnd)) {
-			return index;
-		}
-	}
-
-	// first level as default
-	return LEADERBOARD_RANKS.length - 1;
+export const roundWithDecimals = (value: number, decimals: number): number => {
+	const decimalsCriteria = Math.pow(10, decimals);
+	return Math.round((value + Number.EPSILON) * decimalsCriteria) / decimalsCriteria;
 };