+++ changelog.md
@@ -1,5 +1,10 @@
 # v2 UI Changelog
 
+### Hotfix - 07/27/2021
+
+-   fixed an issue causing geyser unstake tab to not render
+-   fixed an issue causing eth third party reward tokens to not display
+
 ### v2.8.4 - 07/22/2021
 
 -   update boost leaderboard visualization