+++ changelog.md
@@ -1,5 +1,10 @@
 # v2 UI Changelog
 
+### v2.8.5 - 08/05/2021
+
+-   Badger Boost v2 launch
+-   Boost Optimizer release
+
 ### Hotfix - 07/31/2021
 
 -   restored getCurrentBlock calls to fix ibBTC ROI display